Move `ReservationSet` out of `lib.rs`
