+++ crates/execution/ab-riscv-interpreter/src/lib.rs
@@ -576,25 +576,6 @@ where
     }
 }
 
-/// Reservation set used to implement `Zalrsc` extension's `lr`/`sc` instruction pairs.
-///
-/// `lr` places a reservation on an address, and a subsequent `sc` succeeds only if the reservation
-/// is still held for the same address. Regardless of success or failure, executing `sc` always
-/// invalidates the reservation, as does a subsequent `lr`.
-pub const trait ReservationSet<Reg>
-where
-    Reg: [const] Register,
-{
-    /// Returns the address of the currently held reservation, if any
-    fn reservation(&self) -> Option<Reg::Type>;
-
-    /// Place a reservation on `address`, replacing any previously held reservation
-    fn set_reservation(&mut self, address: Reg::Type);
-
-    /// Clear any currently held reservation
-    fn clear_reservation(&mut self);
-}
-
 /// Custom handler for system instructions `ecall` and `ebreak`
 pub const trait SystemInstructionHandler<
     Reg,
