+++ crates/execution/ab-riscv-interpreter/src/prelude.rs
@@ -1,5 +1,6 @@
 //! Re-export of all traits, core types, and instruction helpers
 
+pub use crate::rv32::a::ReservationSet;
 pub use crate::rv32::b::zbb::rv32_zbb_helpers;
 pub use crate::rv32::b::zbc::rv32_zbc_helpers;
 pub use crate::rv32::zce::zcmp::rv32_zcmp_helpers;
@@ -37,6 +38,6 @@ pub use crate::{
     BasicInt, CsrError, Csrs, ExecutableInstruction, ExecutableInstructionCsr,
     ExecutableInstructionOperands, ExecutableInstructionResult, ExecutionError,
     FetchInstructionResult, InstructionFetcher, ProgramCounter, ProgramCounterError, RegisterFile,
-    ReservationSet, Rs1Rs2OperandValues, Rs1Rs2Operands, SystemInstructionHandler, VirtualMemory,
+    Rs1Rs2OperandValues, Rs1Rs2Operands, SystemInstructionHandler, VirtualMemory,
     VirtualMemoryError,
 };
