+++ crates/execution/ab-riscv-interpreter/src/rv32/a.rs
@@ -5,12 +5,30 @@ pub mod zalrsc;
 
 use crate::{
     ExecutableInstruction, ExecutableInstructionCsr, ExecutableInstructionOperands,
-    ExecutableInstructionResult, RegisterFile, ReservationSet, Rs1Rs2OperandValues, Rs1Rs2Operands,
-    VirtualMemory,
+    ExecutableInstructionResult, RegisterFile, Rs1Rs2OperandValues, Rs1Rs2Operands, VirtualMemory,
 };
 use ab_riscv_macros::instruction_execution;
 use ab_riscv_primitives::prelude::*;
 use core::ops::ControlFlow;
+
+/// Reservation set used to implement `Zalrsc` extension's `lr`/`sc` instruction pairs.
+///
+/// `lr` places a reservation on an address, and a subsequent `sc` succeeds only if the reservation
+/// is still held for the same address. Regardless of success or failure, executing `sc` always
+/// invalidates the reservation, as does a subsequent `lr`.
+pub const trait ReservationSet<Reg>
+where
+    Reg: [const] Register,
+{
+    /// Returns the address of the currently held reservation, if any
+    fn reservation(&self) -> Option<Reg::Type>;
+
+    /// Place a reservation on `address`, replacing any previously held reservation
+    fn set_reservation(&mut self, address: Reg::Type);
+
+    /// Clear any currently held reservation
+    fn clear_reservation(&mut self);
+}
 
 #[instruction_execution]
 const impl<Reg> ExecutableInstructionOperands for Rv32AInstruction<Reg> where
