+++ crates/execution/ab-riscv-interpreter/src/rv64/a/zalrsc.rs
@@ -3,10 +3,10 @@
 #[cfg(test)]
 mod tests;
 
+use crate::rv32::a::ReservationSet;
 use crate::{
     ExecutableInstruction, ExecutableInstructionCsr, ExecutableInstructionOperands,
-    ExecutableInstructionResult, RegisterFile, ReservationSet, Rs1Rs2OperandValues, Rs1Rs2Operands,
-    VirtualMemory,
+    ExecutableInstructionResult, RegisterFile, Rs1Rs2OperandValues, Rs1Rs2Operands, VirtualMemory,
 };
 use ab_riscv_macros::instruction_execution;
 use ab_riscv_primitives::prelude::*;
