+++ crates/execution/ab-riscv-interpreter/src/rv64/test_utils.rs
@@ -1,6 +1,7 @@
 extern crate alloc;
 
 use crate::basic::{BasicInterpreterState, BasicRegisters};
+use crate::rv32::a::ReservationSet;
 use crate::v::vector_registers::{
     VectorRegisterFile, VectorRegisters, VectorRegistersBase, VectorRegistersExt,
 };
@@ -8,7 +9,7 @@ use crate::zawrs::WrsHandler;
 use crate::{
     Address, BasicInt, CsrError, Csrs, ExecutableInstruction, ExecutableInstructionCsr,
     ExecutionError, FetchInstructionResult, InstructionFetcher, ProgramCounter,
-    ProgramCounterError, RegisterFile, ReservationSet, Rs1Rs2OperandValues, Rs1Rs2Operands,
+    ProgramCounterError, RegisterFile, Rs1Rs2OperandValues, Rs1Rs2Operands,
     SystemInstructionHandler, VirtualMemory, VirtualMemoryError,
 };
 use ab_riscv_primitives::prelude::*;
